added warning flags

+++ src/ezudp-server.c
@@ -12,7 +12,7 @@
 #include "ezudp-server.h"
 
 int ezudp_server (
-   short port, long addr,
+   uint16_t port, uint32_t addr,
    int (*cb) (socket_t, void *), void *cb_args) {
    struct sockaddr_in si_me;
